backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
